+++ .eslintrc.js
@@ -1,30 +1,33 @@
 module.exports = {
-    "env": {
-        "es6": true,
-        "node": true,
-        "mocha": true
-    },
-    "extends": "eslint:recommended",
-    "parserOptions": {
-        "sourceType": "module"
-    },
-    "rules": {
-        "indent": [
-            "error",
-            2
-        ],
-        "linebreak-style": [
-            "error",
-            "unix"
-        ],
-        "quotes": [
-            "error",
-            "single"
-        ],
-        "semi": [
-            "error",
-            "always"
-        ],
-        "no-console": "off"
-    }
+  env: {
+    es6: true,
+    node: true,
+    mocha: true
+  },
+  extends: [
+    'eslint:recommended',
+  ],
+  parserOptions: {
+    sourceType: 'module',
+    ecmaVersion: 2017,
+  },
+  rules: {
+    indent: [
+      'error',
+      2
+    ],
+    'linebreak-style': [
+      'error',
+      'unix'
+    ],
+    quotes: [
+      'error',
+      'single'
+    ],
+    semi: [
+      'error',
+      'always'
+    ],
+    'no-console': 'off'
+  }
 };