+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.0.0
+
+- Upgrades bluebird and commander dependencies
+- Alpha/Beta verbose logging removed
+
+## 1.0.0-alpha.12, 1.0.0-beta.1
+
+- Supports Storybook feature storyStoreV7
+- Puppeteer upgraded to ^18.2.1
+
 ## 0.24.0
 - Update screener-runner to v0.14.0